+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next'
 import { Graph } from 'types/'
 import { getShortestPath } from 'utils/hooks/dijkstra'
-import { useAuth } from 'utils/hooks/useAuth'
+import { useAuth } from 'utils/hooks/auth/useAuth'
 
 const Home: NextPage = () => {
   const { signIn } = useAuth()
